requirements.txt
tabulate==0.9.0
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.2